+++ tests/testthat/helper-integration-tests-setup-function.R
@@ -451,7 +451,7 @@ setup_and_run_FIMS_without_wrappers <- function(iter_id,
       control = list(eval.max = 10000, iter.max = 10000, trace = 0)
     )
     FIMS::set_fixed(opt$par)
-    fims_finalized <- caa$get_output()
+    fims_finalized <- caa$get_output(do_sd_report = estimation_mode)
   }
 
   # Call report using MLE parameter values, or
@@ -543,7 +543,11 @@ setup_and_run_FIMS_with_wrappers <- function(iter_id,
   clear()
 
   data <- FIMS::FIMSFrame(data1)
-  parameters <- modified_parameters
+  if (tibble::is_tibble(modified_parameters)) {
+    parameters <- modified_parameters
+  } else {
+    parameters <- modified_parameters[[iter_id]]
+  }
 
   # The model will not always run when log_q is very small.
   # We will need to make sure log_q is the true value for deterministic runs but