+++ tests/testthat/test-integration-caa-mle.R
@@ -10,11 +10,11 @@
 # Deterministic test ----
 ## Setup ----
 # Load necessary data for the integration test
-if (!file.exists(test_path("fixtures", "fit_age_length_comp.RDS"))) {
+if (!file.exists(testthat::test_path("fixtures", "fit_age_length_comp.RDS"))) {
   prepare_test_data()
 }
 
-load(test_path("fixtures", "integration_test_data.RData"))
+load(testthat::test_path("fixtures", "integration_test_data.RData"))
 
 # Set the iteration ID to 1 for accessing specific input/output list
 iter_id <- 1
@@ -97,10 +97,11 @@ test_that("catch-at-age model (estimation MLE without wrappers) works with corre
 })
 
 ## Edge handling ----
-test_that("catch-at-age model (estimation MLE without wrappers) returns correct outputs for edge cases", {
+test_that("catch-at-age model (estimation MLE without wrappers) returns correct outputs for missing values", {
   # Introduce a missing value in survey observations for the EM input
   na_value <- -999
   na_index <- 2
+  orig_val <- em_input_list[[iter_id]][["surveyB.obs"]][["survey1"]][na_index]
   em_input_list[[iter_id]][["surveyB.obs"]][["survey1"]][na_index] <- na_value
 
   # Run the FIMS setup and execution function
@@ -129,6 +130,19 @@ test_that("catch-at-age model (estimation MLE without wrappers) returns correct 
   #' @description Test that the maximum gradient is less than or equal to 0.0001 after running FIMS with missing values in survey index observations.
   expect_lte(max_gradient, 0.0001)
 
+  #' @description Test that FIMS runs with only one observation in landings and survey data and the report is not NULL.
+  validate_fims(
+    report = result[["report"]],
+    estimates = result[["sdr_report"]],
+    om_input = om_input_list[[iter_id]],
+    om_output = om_output_list[[iter_id]],
+    em_input = em_input_list[[iter_id]]
+  )
+
+  em_input_list[[iter_id]][["surveyB.obs"]][["survey1"]][na_index] <- orig_val
+})
+
+test_that("catch-at-age model (estimation MLE without wrappers) returns correct outputs for sample size of 1", {
   # Store the original values of the number of landings observations and
   # survey observations
   n.L_original <- om_input_list[[iter_id]][["n.L"]][["fleet1"]]
@@ -137,14 +151,6 @@ test_that("catch-at-age model (estimation MLE without wrappers) returns correct 
   # Set the number of landings observations and survey observations to 1
   om_input_list[[iter_id]][["n.L"]][["fleet1"]] <- 1
   om_input_list[[iter_id]][["n.survey"]][["survey1"]] <- 1
-  on.exit(
-    om_input_list[[iter_id]][["n.L"]][["fleet1"]] <- n.L_original,
-    add = TRUE
-  )
-  on.exit(
-    om_input_list[[iter_id]][["n.survey"]][["survey1"]] <- n.survey_original,
-    add = TRUE
-  )
 
   # Run the FIMS setup and execution function
   result <- setup_and_run_FIMS_without_wrappers(
@@ -163,7 +169,48 @@ test_that("catch-at-age model (estimation MLE without wrappers) returns correct 
     om_output = om_output_list[[iter_id]],
     em_input = em_input_list[[iter_id]]
   )
+
+  om_input_list[[iter_id]][["n.L"]][["fleet1"]] <- n.L_original
+  om_input_list[[iter_id]][["n.survey"]][["survey1"]] <- n.survey_original
 })
 
+test_that("catch-at-age model (estimation MLE without wrappers) returns correct outputs with NA values and sample size of 1", {
+  # Introduce a missing value in survey observations for the EM input
+  na_value <- -999
+  na_index <- 2
+  orig_val <- em_input_list[[iter_id]][["surveyB.obs"]][["survey1"]][na_index]
+  em_input_list[[iter_id]][["surveyB.obs"]][["survey1"]][na_index] <- na_value
+
+  # Store the original values of the number of landings observations and
+  # survey observations
+  n.L_original <- om_input_list[[iter_id]][["n.L"]][["fleet1"]]
+  n.survey_original <- om_input_list[[iter_id]][["n.survey"]][["survey1"]]
+
+  # Set the number of landings observations and survey observations to 1
+  om_input_list[[iter_id]][["n.L"]][["fleet1"]] <- 1
+  om_input_list[[iter_id]][["n.survey"]][["survey1"]] <- 1
+
+  # Run the FIMS setup and execution function
+  result <- setup_and_run_FIMS_without_wrappers(
+    iter_id = iter_id,
+    om_input_list = om_input_list,
+    om_output_list = om_output_list,
+    em_input_list = em_input_list,
+    estimation_mode = TRUE
+  )
+
+  #' @description Test that FIMS runs with only one observation in landings and survey data and the report is not NULL.
+  validate_fims(
+    report = result[["report"]],
+    estimates = result[["sdr_report"]],
+    om_input = om_input_list[[iter_id]],
+    om_output = om_output_list[[iter_id]],
+    em_input = em_input_list[[iter_id]]
+  )
+
+  om_input_list[[iter_id]][["n.L"]][["fleet1"]] <- n.L_original
+  om_input_list[[iter_id]][["n.survey"]][["survey1"]] <- n.survey_original
+  em_input_list[[iter_id]][["surveyB.obs"]][["survey1"]][na_index] <- orig_val
+})
 ## Error handling ----
 # No built-in errors to test.